Restoring hardcoded width/height. New .sh to automate the work.
There was an issue where cached images would not be able to provide image dimensions in time to avoid their values from getting read, which is a huge problem for border rendering and automatic credits display. This restores the hardcoded values and provides a script to automatically fetch them from files. Requires ImageMagick to be installed.